:sparkles: Show / Hide for Rules, Legend & Stats
On mobile, you had to scroll past these sections to get to the upcoming fixtures / results. Making them optional allows mobile users (which let's face it, is everyone but me at this point) can get straight to the action ASAP.

If anyone needs a refresher on how things work or wants to see how they're doing in terms of stats, they are free to do so!